Replace reach_error (or equivalent) with assert(0)
Previously, all unreach_call benchmarks would call reach_error() which
would contain assert(0). 2LS and CBMC relied on this fact to check the
property. However, as per the rules, assert(0) may not necesarilly be
present and just calling a function specified in the property file
suffices.

To overcome this, parse the function which causes a fail from the
property file and replace its body with assert(0) using goto-instrument.

Signed-off-by: František Nečas <frantisek.necas@protonmail.com>